pyproject.toml
[project]
name = "SAFTEC"
version = "1.0.0"
description = "Application for SAFTEC"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fleting==1.1.34",
    "black>=26.3.1",
    "openpyxl>=3.1.5",
    "pandas>=3.0.1",
    "playwright>=1.58.0",
    "httpx>=0.28.1",
    "packaging>=26.1",
]

[tool.flet]
product = "Application for SAFTEC"
artifact = "SAFTEC-app"
company = "DASF"
org = "sesab.dasf"
copyright = "Copyright (C) 2026 Bernardo Nogueira"
build_number = 1

[tool.flet.windows]
resources = ["ms-playwright"]
